+++ R/app-driver-expect-values.R
@@ -4,9 +4,9 @@
 # `app_expect_values()` retrieves the JSON text format of the shiny values.
 #   * All images are hashed within the text content.
 
-
 app_get_single_ioe <- function(
-  self, private,
+  self,
+  private,
   ...,
   input = missing_arg(),
   output = missing_arg(),
@@ -23,22 +23,28 @@ app_get_single_ioe <- function(
   export_is_provided <- !rlang::is_missing(export)
 
   if (sum(input_is_provided, output_is_provided, export_is_provided) != 1) {
-    app_abort(self, private, "Must specify only one of `input`, `output`, `export`")
+    app_abort(
+      self,
+      private,
+      "Must specify only one of `input`, `output`, `export`"
+    )
   }
 
-  if (input_is_provided) ckm8_assert_single_string(input)
-  else if (output_is_provided) ckm8_assert_single_string(output)
-  else if (export_is_provided) ckm8_assert_single_string(export)
-  else app_abort(self, private, "Missing ioe type", .internal = TRUE)
+  if (input_is_provided) ckm8_assert_single_string(input) else if (
+    output_is_provided
+  )
+    ckm8_assert_single_string(output) else if (export_is_provided)
+    ckm8_assert_single_string(export) else
+    app_abort(self, private, "Missing ioe type", .internal = TRUE)
 
   type <-
-    if (input_is_provided) "input"
-    else if (output_is_provided) "output"
-    else if (export_is_provided) "export"
+    if (input_is_provided) "input" else if (output_is_provided)
+      "output" else if (export_is_provided) "export"
   name <-
-    if (input_is_provided) input
-    else if (output_is_provided) output
-    else if (export_is_provided) export
+    if (input_is_provided) input else if (output_is_provided) output else if (
+      export_is_provided
+    )
+      export
 
   list(
     input = input,
@@ -50,19 +56,27 @@ app_get_single_ioe <- function(
 }
 
 app_get_value <- function(
-  self, private,
+  self,
+  private,
   ...,
-  input = missing_arg(), output = missing_arg(), export = missing_arg(),
+  input = missing_arg(),
+  output = missing_arg(),
+  export = missing_arg(),
   hash_images = FALSE
 ) {
   ioe <- app_get_single_ioe(
-    self, private,
-    input = input, output = output, export = export
+    self,
+    private,
+    input = input,
+    output = output,
+    export = export
   )
 
   # Call `app_get_values()` to get the RDS format of the shiny values.
   ret <- self$get_values(
-    input = ioe$input, output = ioe$output, export = ioe$export,
+    input = ioe$input,
+    output = ioe$output,
+    export = ioe$export,
     hash_images = hash_images
   )
 
@@ -73,9 +87,12 @@ app_get_value <- function(
 
 # Note: This queries method the Shiny server
 app_get_values <- function(
-  self, private,
+  self,
+  private,
   ...,
-  input = missing_arg(), output = missing_arg(), export = missing_arg(),
+  input = missing_arg(),
+  output = missing_arg(),
+  export = missing_arg(),
   hash_images = FALSE
 ) {
   ckm8_assert_app_driver(self, private)
@@ -85,13 +102,21 @@ app_get_values <- function(
   self$log_message("Getting all values")
   "!DEBUG app_get_values()"
 
-  url <- app_get_shiny_test_url(self, private, input, output, export, format = "rds")
+  url <- app_get_shiny_test_url(
+    self,
+    private,
+    input,
+    output,
+    export,
+    format = "rds"
+  )
 
   # Ask Shiny for info
   cur_env <- rlang::current_env()
   req <- app_httr_get(self, private, url, fn_404 = function(req) {
     app_abort(
-      self, private,
+      self,
+      private,
       c(
         paste0("Shiny server returned 404 for values URL: ", url),
         "i" = "Is `shiny::runApp(test.mode = TRUE)` enabled?"
@@ -144,14 +169,14 @@ app_get_values <- function(
 
 
 app_expect_values <- function(
-  self, private,
+  self,
+  private,
   ...,
   input = missing_arg(),
   output = missing_arg(),
   export = missing_arg(),
   name = NULL,
-  screenshot_args = missing_arg(),
-  cran = FALSE
+  screenshot_args = missing_arg()
 ) {
   ckm8_assert_app_driver(self, private)
   rlang::check_dots_empty()
@@ -160,17 +185,28 @@ app_expect_values <- function(
   "!DEBUG app_expect_values()"
 
   screenshot_args_ <- default_screenshot_args(
-    maybe_missing_value(screenshot_args, private$default_expect_values_screenshot_args)
+    maybe_missing_value(
+      screenshot_args,
+      private$default_expect_values_screenshot_args
+    )
   )
 
   json_path <- app_next_temp_snapshot_path(self, private, name, "json")
 
-  url <- app_get_shiny_test_url(self, private, input, output, export, format = "json")
+  url <- app_get_shiny_test_url(
+    self,
+    private,
+    input,
+    output,
+    export,
+    format = "json"
+  )
   # Ask Shiny for info
   cur_env <- rlang::current_env()
   req <- app_httr_get(self, private, url, fn_404 = function(req) {
     app_abort(
-      self, private,
+      self,
+      private,
       c(
         paste0("Shiny server returned 404 for values URL: ", url),
         "i" = "Is `shiny::runApp(test.mode = TRUE)` enabled?"
@@ -189,7 +225,6 @@ app_expect_values <- function(
   # Take a screenshot for debugging purposes
   # If `screenshot_args_` is anything other than `FALSE`...
   if (!is_false(screenshot_args_)) {
-
     # If a single output is used, then zoom in on the output
     # TODO-future; Should we do this for `inputs` as well! Works nicely for `outputs`!
     if (length(output) == 1 && is.character(output)) {
@@ -199,8 +234,10 @@ app_expect_values <- function(
         screenshot_args_
       )
     }
-    withCallingHandlers( # swallow expectation
-      { # nolint
+    withCallingHandlers(
+      # swallow expectation
+      {
+        # nolint
         # `NAME.json` -> `NAME_.png`; `NAME_.new.png`
         png_path <-
           fs::path_ext_set(
@@ -211,7 +248,8 @@ app_expect_values <- function(
         # Skip the variant check in `$expect_snapshot()`
         # Leverage testthat snapshot logic, but muffle any expectation output
         app_expect_screenshot(
-          self, private,
+          self,
+          private,
           name = png_path,
           screenshot_args = screenshot_args_
         )
@@ -230,9 +268,9 @@ app_expect_values <- function(
   on.exit(unlink(json_path), add = TRUE)
   # Assert json file contents
   app__expect_snapshot_file(
-    self, private,
+    self,
+    private,
     json_path,
-    cran = cran,
     compare = testthat::compare_file_text
   )
 
@@ -241,7 +279,8 @@ app_expect_values <- function(
 
 
 app_get_shiny_test_url <- function(
-  self, private,
+  self,
+  private,
   input = missing_arg(),
   output = missing_arg(),
   export = missing_arg(),
@@ -263,12 +302,8 @@ app_get_shiny_test_url <- function(
   }
 
   q_string <- function(group, value) {
-    if (isTRUE(value))
-      paste0(group, "=1")
-    else if (is.character(value))
-      paste0(group, "=", paste(value, collapse = ","))
-    else
-      ""
+    if (isTRUE(value)) paste0(group, "=1") else if (is.character(value))
+      paste0(group, "=", paste(value, collapse = ",")) else ""
   }
   paste(
     private$shiny_test_url,
@@ -313,9 +348,12 @@ hash_snapshot_image_data <- function(
   }
 
   # Image data indices
-  image_start_idx <- as.integer(attr2(image_offsets, "capture.start")[, "img_data"])
+  image_start_idx <- as.integer(attr2(image_offsets, "capture.start")[,
+    "img_data"
+  ])
   image_stop_idx <- image_start_idx +
-    as.integer(attr2(image_offsets, "capture.length")[, "img_data"]) - 1
+    as.integer(attr2(image_offsets, "capture.length")[, "img_data"]) -
+    1
 
   # Text (non-image) data indices
   text_start_idx <- c(
@@ -330,17 +368,20 @@ hash_snapshot_image_data <- function(
 
   # Get the strings representing image data, and all the other stuff
   image_data <- substring(data, image_start_idx, image_stop_idx)
-  text_data  <- substring(data, text_start_idx,  text_stop_idx)
+  text_data <- substring(data, text_start_idx, text_stop_idx)
   # Hash the images
   image_hashes <- vapply(image_data, FUN.VALUE = "", function(dat) {
-    tryCatch({
-      image_data <- jsonlite::base64_dec(dat)
-      rlang::hash(
-        image_data
-      )
-    }, error = function(e) {
-      "Error hashing image data"
-    })
+    tryCatch(
+      {
+        image_data <- jsonlite::base64_dec(dat)
+        rlang::hash(
+          image_data
+        )
+      },
+      error = function(e) {
+        "Error hashing image data"
+      }
+    )
   })
 
   image_hashes <- paste0("[image data hash: ", image_hashes, "]")