+++ man-roxygen/app-cran.R
@@ -1,3 +1,2 @@
-#' @param cran Should these expectations be verified on CRAN? By default,
-#'        they are not, because snapshot tests tend to be fragile
-#'        because they often rely on minor details of dependencies.
+#' @param cran Deprecated. With `AppDriver` never testing on CRAN,
+#'             this parameter no longer has any effect.