+++ NEWS.md
@@ -1,5 +1,9 @@
 # shinytest2 (development version)
 
+## Breaking changes
+
+* From a request from CRAN, creating an `AppDriver` (which uses `{chromote}`) during testing will cause the test to skip during CRAN testing. `{shinytest2}` performs this by internally calling `testthat::skip_on_cran()`.
+
 ## Bug / Improvements
 
 * Add support for `$click()`ing `{bslib}`'s `input_task_button()` (#389).
@@ -8,7 +12,7 @@
 
 * The `threshold` and `kernel_size` default values of the `AppDriver$expect_screenshot()` method are now configurable via two new global options: `shinytest2.compare_screenshot.threshold` and `shinytest2.compare_screenshot.kernel_size` (#401)
 
-* `{shinytest2}` now imports `{cli}` and no longer imports `{crayon}` (@olivroy, #399). 
+* `{shinytest2}` now imports `{cli}` and no longer imports `{crayon}` (@olivroy, #399).
 
 # shinytest2 0.3.2